requirements.txt
pandas
scikit-learn
requests
joblib
pyyaml
